server/requirements.txt
fastapi>=0.115
uvicorn[standard]>=0.34
aiosqlite>=0.20
python-jose[cryptography]>=3.3
python-multipart>=0.0.9
google-auth>=2.29
httpx>=0.27
